delete `AzResourceBase` and merge into `AzResource`
