+++ azure-toolkit-libs/azure-toolkit-common-lib/src/main/java/com/microsoft/azure/toolkit/lib/common/model/Startable.java
@@ -5,7 +5,7 @@
 
 package com.microsoft.azure.toolkit.lib.common.model;
 
-public interface Startable extends AzResourceBase {
+public interface Startable extends AzResource {
 
     void start();
