+++ azure-toolkit-libs/azure-toolkit-cosmos-lib/src/main/java/com/microsoft/azure/toolkit/lib/cosmos/ICosmosDatabase.java
@@ -1,9 +1,9 @@
 package com.microsoft.azure.toolkit.lib.cosmos;
 
-import com.microsoft.azure.toolkit.lib.common.model.AzResourceBase;
+import com.microsoft.azure.toolkit.lib.common.model.AzResource;
 
 import java.util.List;
 
-public interface ICosmosDatabase extends AzResourceBase {
+public interface ICosmosDatabase extends AzResource {
     List<? extends ICosmosCollection> listCollection();
 }
